requirement.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
jsonschema==4.21.1
motor==4.5.0
groq==0.3.0
